Revert ai slop
