+++ comp/dogstatsd/server/intern_telemetry.go
@@ -19,7 +19,8 @@ type stringInternerTelemetry struct {
 }
 
 type stringInternerInstanceTelemetry struct {
-	enabled bool
+	enabled  bool
+	curBytes int
 
 	resets               telemetry.SimpleCounter
 	size                 telemetry.SimpleGauge
@@ -72,5 +73,26 @@ func (s *stringInternerTelemetry) PrepareForID(id string) *stringInternerInstanc
 func (si *stringInternerInstanceTelemetry) Hit() {
 	if si.enabled {
 		si.hits.Inc()
+	}
+}
+
+// Reset increments the reset counter and updates the size and bytes gauges.
+func (si *stringInternerInstanceTelemetry) Reset(length int) {
+	if si.enabled {
+		si.resets.Inc()
+		si.bytes.Sub(float64(si.curBytes))
+		si.size.Sub(float64(length))
+		si.curBytes = 0
+	}
+}
+
+// Miss increments the miss counter and updates the size and bytes gauges.
+func (si *stringInternerInstanceTelemetry) Miss(length int) {
+	if si.enabled {
+		si.miss.Inc()
+		si.size.Inc()
+		si.bytes.Add(float64(length))
+		si.globaltlmSIRStrBytes.Observe(float64(length))
+		si.curBytes += length
 	}
 }
